Talk_with_Google_Gemini/requirements.txt
streamlit 
google-generativeai
python-dotenv

# wordle env pe chala le
